Move `UTxOIndex` type to `cardano-coin-selection` library.
